Fix double pattern search issue
"I have however discovered an issue whereby wrong results are returned.

I have used the multi-pattern solver to solver 2 words at the same time.
The patterns are:
12..S
S.E21

The issue with the results is that they are returning lines where the first word contains an E, but the E has been specified on the second word and not included in the first, therefore the first word cannot contain ‘E’

A couple of examples from the results MAZES:STEAM, TALES: SWEAT.  I also
had a result ONE’S:STENO"

This is because we exclude fixed letters in the first word from the
second word results, but not vice-versa.

Quickfix is to run the double pattern search twice with the words
flipped around.